+++ src/main/Oidc4vciErrors.ts
