+++ src/main/functions/HttpUtils.ts
@@ -1,8 +1,8 @@
 import { fetch } from 'cross-fetch';
 
-import { Request } from '../types';
+import { CredentialRequest } from '../types';
 
-export async function postWithBearerToken(url: string, body: Request, bearerToken: string): Promise<Response> {
+export async function postWithBearerToken(url: string, body: CredentialRequest, bearerToken: string): Promise<Response> {
   try {
     const response = await fetch(url, {
       method: 'POST',
