+++ src/main/types/VCIssuance.types.ts
@@ -1,16 +1,27 @@
 import { W3CVerifiableCredential } from '@sphereon/ssi-types';
 import { ClaimFormat } from '@sphereon/ssi-types/src/types/vc';
 
+export enum AuthzFlowType {
+  AUTHORIZATION_CODE_FLOW = 'Authorization Code Flow',
+  PRE_AUTHORIZED_CODE_FLOW = 'Pre-Authorized Code Flow',
+}
+
+// eslint-disable-next-line @typescript-eslint/no-namespace
+export namespace AuthzFlowType {
+  export function valueOf(request: IssuanceInitiationRequestPayload) {
+    if (request.pre_authorized_code) {
+      return AuthzFlowType.PRE_AUTHORIZED_CODE_FLOW;
+    }
+    return AuthzFlowType.AUTHORIZATION_CODE_FLOW;
+  }
+}
+
 export interface CredentialRequest {
   //TODO: handling list is out of scope for now
   type: string | string[];
   //TODO: handling list is out of scope for now
   format: ClaimFormat | ClaimFormat[];
   proof: ProofOfPossession;
-}
-
-export enum ProofType {
-  JWT = 'jwt',
 }
 
 export interface CredentialResponse {
@@ -34,10 +45,20 @@ export enum CredentialResponseErrorCode {
   INVALID_CREDENTIAL = 'invalid_credential',
 }
 
+export interface IssuanceInitiationRequestPayload {
+  issuer: string; //(url) REQUIRED The issuer URL of the Credential issuer, the Wallet is requested to obtain one or more Credentials from.
+  credential_type: string[] | string; //(url) REQUIRED A JSON string denoting the type of the Credential the Wallet shall request
+  pre_authorized_code?: string; //CONDITIONAL he code representing the issuer's authorization for the Wallet to obtain Credentials of a certain type. This code MUST be short lived and single-use. MUST be present in a pre-authorized code flow.
+  user_pin_required?: boolean; //OPTIONAL Boolean value specifying whether the issuer expects presentation of a user PIN along with the Token Request in a pre-authorized code flow. Default is false.
+  op_state?: string; //(JWT) OPTIONAL String value created by the Credential Issuer and opaque to the Wallet that is used to bind the sub-sequent authentication request with the Credential Issuer to a context set up during previous steps
+}
+
+export enum ProofType {
+  JWT = 'jwt',
+}
+
 export interface ProofOfPossession {
   proof_type: ProofType;
   jwt: string;
   [x: string]: any;
 }
-
-export type Request = CredentialRequest;
