--- src/types/AuthzFlowType.ts
@@ -1,16 +1,0 @@
-import { IssuanceInitiationRequestPayload } from './IssuanceInitiationRequestTypes';
-
-export enum AuthzFlowType {
-  AUTHORIZATION_CODE_FLOW = 'Authorization Code Flow',
-  PRE_AUTHORIZED_CODE_FLOW = 'Pre-Authorized Code Flow',
-}
-
-// eslint-disable-next-line @typescript-eslint/no-namespace
-export namespace AuthzFlowType {
-  export function valueOf(request: IssuanceInitiationRequestPayload) {
-    if (request.pre_authorized_code) {
-      return AuthzFlowType.PRE_AUTHORIZED_CODE_FLOW;
-    }
-    return AuthzFlowType.AUTHORIZATION_CODE_FLOW;
-  }
-}
